requirements.txt
Flask==3.1.0
google-generativeai==0.8.5
flask-cors==5.0.1
google-genai==1.14.0 # for temp.py
python-dotenv==1.1.0 #for loading .env variable
fastapi==0.115.12
uvicorn==0.34.2
